# pkg/api/static/api-spec.yaml
openapi: 3.0.0
info:
  title: DNSPulse ConfigHUB API
  version: 1.0.0
  description: DNSPulse Configuration Management REST API
servers:
  - url: http://example.com/api
    description: Production server
paths:
...
  /api/config/general/segment/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: Name of the segment
        schema:
          type: string
    get:
      tags:
        - Segment Configuration
      summary: Get a specific segment configuration
      description: Retrieve the configuration for the specified segment
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SegmentConfigsStruct'
    post:
      tags:
        - Segment Configuration
      summary: Create a new specific segment configuration
      description: Add a new specific segment configuration to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SegmentConfigsStruct'
      responses:
        '201':
          description: Specific segment configuration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SegmentConfigsStruct'
    put:
      tags:
        - Segment Configuration
      summary: Replace a specific segment configuration
      description: Replace the entire configuration for the specified segment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SegmentConfigsStruct'
      responses:
        '200':
          description: Specific segment configuration replaced successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SegmentConfigsStruct'
    patch:
      tags:
        - Segment Configuration
      summary: Modify a specific segment configuration
      description: Apply partial modifications to the configuration of the specified segment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SegmentConfigsStruct'
      responses:
        '200':
          description: Specific segment configuration modified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      tags:
        - Segment Configuration
      summary: Delete a specific segment configuration
      description: Remove the configuration for the specified segment from the system
      responses:
        '200':
          description: Specific segment configuration deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SegmentConfigsStruct:
      type: object
      properties:
        Name:
          type: string
          description: Name of the segment
        Path:
          type: string
          description: Filesystem path associated with the segment configuration
      required:
        - Name
        - Path
      example:
        Name: "PrimaryDataCenter"
        Path: "/config/segments/primary.yaml"
...
security:
  - BearerAuth: []